[RENAME] reviewPostListGET to reviewPostListPOST
