+++ functions/api/routes/reviewPost/index.js
@@ -8,7 +8,7 @@ router.delete("/:postId", checkUser, require("./reviewPostDELETE"));
 router.get("/:postId", checkUser, require("./reviewPostDetailGET"));
 router.get("/tag/list", checkUser, require("./reviewPostTagListGET"));
 router.get("/background-image/list", checkUser, require("./reviewPostBackgroundImageListGET"));
-router.post("/list", checkUser, require("./reviewPostListGET"));
+router.post("/list", checkUser, require("./reviewPostListPOST"));
 router.put("/:postId", checkUser, require("./reviewPostPUT"));
 
 module.exports = router;
